+++ modules/web/js/ballerina/ast/ast-factory.js
@@ -1,4 +1,4 @@
-/**
+  /**
  * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
