+++ modules/web/js/ballerina/ast/expressions/type-conversion-expression.js
@@ -83,6 +83,10 @@ class TypeConversionExpression extends Expression {
                 + this.getWSRegion(2) + this.children[0].getExpressionString();
         return expString;
     }
+
+    getRightExpression() {
+        return this.children[0];
+    }
 }
 
 export default TypeConversionExpression;
